cleanup

+++ src/Mocha/src/Mocha.Transport.RabbitMQ/RabbitMQReceiveEndpoint.cs
@@ -28,10 +28,11 @@ public sealed class RabbitMQReceiveEndpoint(RabbitMQMessagingTransport transport
         }
 
         _maxPrefetch = configuration.MaxPrefetch;
-        _consumerDispatchConcurrency = (ushort)Math.Clamp(
-            configuration.MaxConcurrency ?? ReceiveEndpointConfiguration.Defaults.MaxConcurrency,
-            1,
-            (int)ushort.MaxValue);
+        _consumerDispatchConcurrency = (ushort)
+            Math.Clamp(
+                configuration.MaxConcurrency ?? ReceiveEndpointConfiguration.Defaults.MaxConcurrency,
+                1,
+                ushort.MaxValue);
     }
 
     protected override void OnComplete(
